Add placeholder files for tests
